requirements.txt
pyyaml
requests
numpy
web3
